requirements.txt
anyio==3.6.2
APScheduler==3.10.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.95.1
h11==0.14.0
httptools==0.5.0
idna==3.4
loguru==0.6.0
msgpack==1.0.5
multidict==6.0.4
nonebot-adapter-onebot==2.2.3
nonebot-plugin-apscheduler==0.2.0
nonebot-plugin-firexN==0.1.7
nonebot2==2.0.0rc4
pydantic==1.10.7
pygtrie==2.5.0
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.30.0
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==2.0.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
